Show text of selected non-truthy items
When adding a SelectOption Interface with an item whose value is non-truthy (for example 0),
this option's text should be shown when it is selected. If value is set to an option that is not in
the list of options, getItemByValue will return undefined, which will be converted to an empty
string, as it was before.